requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-api-python-client==2.179.0
requests==2.31.0
APScheduler==3.11.0
Pillow==10.2.0
authlib==1.6.3
